requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
langchain
langchain_google_genai
python-dotenv